Extract logic out of CaptureActivity.
