+++ zxing-android-embedded/res/layout/zxing_capture.xml
@@ -17,22 +17,9 @@
 <merge xmlns:android="http://schemas.android.com/apk/res/android"
        xmlns:tools="http://schemas.android.com/tools">
 
-    <com.journeyapps.barcodescanner.BarcodeView
+    <com.journeyapps.barcodescanner.DefaultBarcodeScannerView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/zxing_barcode_surface"/>
+        android:id="@+id/zxing_barcode_scanner"/>
 
-    <com.journeyapps.barcodescanner.ViewfinderView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/zxing_viewfinder_view"/>
-
-
-    <TextView android:id="@+id/zxing_status_view"
-              android:layout_width="wrap_content"
-              android:layout_height="wrap_content"
-              android:layout_gravity="bottom|center_horizontal"
-              android:background="@color/zxing_transparent"
-              android:text="@string/zxing_msg_default_status"
-              android:textColor="@color/zxing_status_text"/>
 </merge>
